AI-Service/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
pytz
python-dotenv
